functions/load-data/bq-load-file/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-auth==2.12.0
google-cloud-appengine-logging==1.1.5
google-cloud-audit-log==0.2.4
google-cloud-bigquery==3.3.3
google-cloud-bigquery-storage==2.16.1
google-cloud-core==2.3.2
google-cloud-datastore==2.8.2
google-cloud-logging==3.2.4
google-cloud-storage==2.5.0
google-cloud-tasks==2.10.3
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-status==1.49.1
idna==3.4
numpy==1.23.3
packaging==21.3
proto-plus==1.22.1
protobuf==4.21.7
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.12